chore: request tracking permission and use SPM
