--- TrinityPlayer.xcframework/ios-arm64/TrinityPlayer.framework/Modules/module.modulemap
@@ -1,11 +1,0 @@
-framework module TrinityPlayer {
-  umbrella header "TrinityPlayer.h"
-  export *
-
-  module * { export * }
-}
-
-module TrinityPlayer.Swift {
-  header "TrinityPlayer-Swift.h"
-  requires objc
-}
