--- TrinityPlayer.xcframework/ios-arm64_x86_64-simulator/TrinityPlayer.framework/Headers/TrinityPlayer.h
@@ -1,18 +1,0 @@
-//
-//  TrinityPlayer.h
-//  TrinityPlayer
-//
-//  Created by Kuznetsov Roman on 1/4/21.
-//
-
-#import <Foundation/Foundation.h>
-
-//! Project version number for TrinityPlayer.
-FOUNDATION_EXPORT double TrinityPlayerVersionNumber;
-
-//! Project version string for TrinityPlayer.
-FOUNDATION_EXPORT const unsigned char TrinityPlayerVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <TrinityPlayer/PublicHeader.h>
-
-
