--- TrinityPlayer.xcframework/ios-arm64_x86_64-simulator/TrinityPlayer.framework/Modules/module.modulemap
@@ -1,11 +1,0 @@
-framework module TrinityPlayer {
-  umbrella header "TrinityPlayer.h"
-  export *
-
-  module * { export * }
-}
-
-module TrinityPlayer.Swift {
-  header "TrinityPlayer-Swift.h"
-  requires objc
-}
