--- TrinityPlayer.xcframework/ios-arm64_x86_64-simulator/TrinityPlayer.framework/_CodeSignature/CodeResources
@@ -1,278 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>files</key>
-	<dict>
-		<key>Headers/TrinityPlayer-Swift.h</key>
-		<data>
-		jXL2I4OUcfMjLBoMiwO0wCjKGpU=
-		</data>
-		<key>Headers/TrinityPlayer.h</key>
-		<data>
-		mimQ8Oyq37MlpTVizR+Jlwy8rXQ=
-		</data>
-		<key>Info.plist</key>
-		<data>
-		2qm8rzSKNh3zl6w7nkk1ZoiHz5Y=
-		</data>
-		<key>Modules/TrinityPlayer.swiftmodule/arm64-apple-ios-simulator.abi.json</key>
-		<data>
-		2nXQxg6MdUqwTdqUxGDOxaff4Ws=
-		</data>
-		<key>Modules/TrinityPlayer.swiftmodule/arm64-apple-ios-simulator.private.swiftinterface</key>
-		<data>
-		KP7r+hR1XByreu3deQkDtkTCmd4=
-		</data>
-		<key>Modules/TrinityPlayer.swiftmodule/arm64-apple-ios-simulator.swiftdoc</key>
-		<data>
-		ukscOXxwrGn9EUmC0q3dmT92Ezs=
-		</data>
-		<key>Modules/TrinityPlayer.swiftmodule/arm64-apple-ios-simulator.swiftinterface</key>
-		<data>
-		KP7r+hR1XByreu3deQkDtkTCmd4=
-		</data>
-		<key>Modules/TrinityPlayer.swiftmodule/arm64-apple-ios-simulator.swiftmodule</key>
-		<data>
-		AYz/5bCCED3pU6j/Yv1KdD3WM1Q=
-		</data>
-		<key>Modules/TrinityPlayer.swiftmodule/x86_64-apple-ios-simulator.abi.json</key>
-		<data>
-		2nXQxg6MdUqwTdqUxGDOxaff4Ws=
-		</data>
-		<key>Modules/TrinityPlayer.swiftmodule/x86_64-apple-ios-simulator.private.swiftinterface</key>
-		<data>
-		x4DYKieEoJ71y/4d6wdIVlgPd1E=
-		</data>
-		<key>Modules/TrinityPlayer.swiftmodule/x86_64-apple-ios-simulator.swiftdoc</key>
-		<data>
-		NJiCqyzd1IzNnxqWkflZKebbAFo=
-		</data>
-		<key>Modules/TrinityPlayer.swiftmodule/x86_64-apple-ios-simulator.swiftinterface</key>
-		<data>
-		x4DYKieEoJ71y/4d6wdIVlgPd1E=
-		</data>
-		<key>Modules/TrinityPlayer.swiftmodule/x86_64-apple-ios-simulator.swiftmodule</key>
-		<data>
-		HWVdpYKZbtiPmAo8FqZOEcWdGYc=
-		</data>
-		<key>Modules/module.modulemap</key>
-		<data>
-		IneMft6UUySMouiBv1GNZ/NUE3o=
-		</data>
-		<key>TrinityPlayerStoryboard.storyboardc/Go4-qm-EiM-view-6OK-cw-yGm.nib</key>
-		<data>
-		P+7biRNHoN2f7lz5auZZ/rJ6WwM=
-		</data>
-		<key>TrinityPlayerStoryboard.storyboardc/Info.plist</key>
-		<data>
-		OA3l+tIjA79MBQZRCMuM9mJ1ZpY=
-		</data>
-		<key>TrinityPlayerStoryboard.storyboardc/TrinityPlayerViewController.nib</key>
-		<data>
-		2IcqCbsxqp49o+ST3DEuS6hBhhI=
-		</data>
-	</dict>
-	<key>files2</key>
-	<dict>
-		<key>Headers/TrinityPlayer-Swift.h</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			kc+0vrlkVKPqUc5eE8BzSXodazjLJMkmA0K0aoXPkbA=
-			</data>
-		</dict>
-		<key>Headers/TrinityPlayer.h</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			u13GQooTJvrICV/7QEnHtcNBzJxaT7Y4/aNOS1ja+0w=
-			</data>
-		</dict>
-		<key>Modules/TrinityPlayer.swiftmodule/arm64-apple-ios-simulator.abi.json</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			B0YF/IcWV/9HlChiHtcK5WYYd4ptL7VK5pRPHqqTKLY=
-			</data>
-		</dict>
-		<key>Modules/TrinityPlayer.swiftmodule/arm64-apple-ios-simulator.private.swiftinterface</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			AqyOQgGrictApvNxjadxVfkf+290TxxHtXD9OHdMVKo=
-			</data>
-		</dict>
-		<key>Modules/TrinityPlayer.swiftmodule/arm64-apple-ios-simulator.swiftdoc</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			sRWna67VsvXIzaH9olNcJs5z16bpHkfDXRbxyrn10Gk=
-			</data>
-		</dict>
-		<key>Modules/TrinityPlayer.swiftmodule/arm64-apple-ios-simulator.swiftinterface</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			AqyOQgGrictApvNxjadxVfkf+290TxxHtXD9OHdMVKo=
-			</data>
-		</dict>
-		<key>Modules/TrinityPlayer.swiftmodule/arm64-apple-ios-simulator.swiftmodule</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			KSNuWaiN+Klc88531g7r0bPrLNAER6c+fUKeuf+Oioo=
-			</data>
-		</dict>
-		<key>Modules/TrinityPlayer.swiftmodule/x86_64-apple-ios-simulator.abi.json</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			B0YF/IcWV/9HlChiHtcK5WYYd4ptL7VK5pRPHqqTKLY=
-			</data>
-		</dict>
-		<key>Modules/TrinityPlayer.swiftmodule/x86_64-apple-ios-simulator.private.swiftinterface</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			JkxjKqvqJ/fyy7PZmHMLIBpI/8lxE4GUogP3v5d+7Z0=
-			</data>
-		</dict>
-		<key>Modules/TrinityPlayer.swiftmodule/x86_64-apple-ios-simulator.swiftdoc</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			fLFjSBLSyXP3vbAtC6EJdR7pGgJTi0sE9oyIjKAhHNM=
-			</data>
-		</dict>
-		<key>Modules/TrinityPlayer.swiftmodule/x86_64-apple-ios-simulator.swiftinterface</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			JkxjKqvqJ/fyy7PZmHMLIBpI/8lxE4GUogP3v5d+7Z0=
-			</data>
-		</dict>
-		<key>Modules/TrinityPlayer.swiftmodule/x86_64-apple-ios-simulator.swiftmodule</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			U89zFODp3e0NTwoTCPpX3+GPRAjm6EyMjyvr2fWfTmQ=
-			</data>
-		</dict>
-		<key>Modules/module.modulemap</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			8f1RTUOtQ7d/UzI/u+mdCrl6rUR2mG1J6S/sOFAeCmE=
-			</data>
-		</dict>
-		<key>TrinityPlayerStoryboard.storyboardc/Go4-qm-EiM-view-6OK-cw-yGm.nib</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			csOKtftHYL8rSYX1D7uZAxEk+QLzMYcM70tnd57rdTI=
-			</data>
-		</dict>
-		<key>TrinityPlayerStoryboard.storyboardc/Info.plist</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			FkblOT3TOjywKMy5pNnbCaeBYVG0EPMVaUCVaQlTu2w=
-			</data>
-		</dict>
-		<key>TrinityPlayerStoryboard.storyboardc/TrinityPlayerViewController.nib</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			+7RFPpDEs7aZI+v6qF0KHK5aAmNb6IaPysL33r9hyiU=
-			</data>
-		</dict>
-	</dict>
-	<key>rules</key>
-	<dict>
-		<key>^.*</key>
-		<true/>
-		<key>^.*\.lproj/</key>
-		<dict>
-			<key>optional</key>
-			<true/>
-			<key>weight</key>
-			<real>1000</real>
-		</dict>
-		<key>^.*\.lproj/locversion.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>1100</real>
-		</dict>
-		<key>^Base\.lproj/</key>
-		<dict>
-			<key>weight</key>
-			<real>1010</real>
-		</dict>
-		<key>^version.plist$</key>
-		<true/>
-	</dict>
-	<key>rules2</key>
-	<dict>
-		<key>.*\.dSYM($|/)</key>
-		<dict>
-			<key>weight</key>
-			<real>11</real>
-		</dict>
-		<key>^(.*/)?\.DS_Store$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>2000</real>
-		</dict>
-		<key>^.*</key>
-		<true/>
-		<key>^.*\.lproj/</key>
-		<dict>
-			<key>optional</key>
-			<true/>
-			<key>weight</key>
-			<real>1000</real>
-		</dict>
-		<key>^.*\.lproj/locversion.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>1100</real>
-		</dict>
-		<key>^Base\.lproj/</key>
-		<dict>
-			<key>weight</key>
-			<real>1010</real>
-		</dict>
-		<key>^Info\.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^PkgInfo$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^embedded\.provisionprofile$</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^version\.plist$</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-	</dict>
-</dict>
-</plist>
